Move some code around to make it cleaner
